rag_application/frontend/requirements.txt
streamlit
chromadb
google-genai
